Component Composition: The special children Prop
