+++ src/components/CoreConcept.jsx
@@ -1,0 +1,13 @@
+import './CoreConcept.css'
+
+function CoreConcept({ title, description, image }) {
+    return (
+        <li>
+            <img src={image} alt={title} />
+            <h3>{title}</h3>
+            <p>{description}</p>
+        </li>
+    )
+}
+
+export default CoreConcept;
