+++ src/components/Header.jsx
@@ -1,0 +1,24 @@
+import reactImg from "../assets/react-core-concepts.png";
+import './Header.css'
+
+const reactDesc = ["Fundamental", "Crucial", "Core"];
+
+function genRandInt(max) {
+    return Math.floor(Math.random() * (max + 1));
+}
+
+function Header() {
+    const description = reactDesc[genRandInt(2)];
+    return (
+        <header>
+            <img src={reactImg} alt="Stylized atom" />
+            <h1>React Essentials</h1>
+            <p>
+                {description} React concepts you will need for almost any app you are
+                going to build!
+            </p>
+        </header>
+    )
+}
+
+export default Header;
