+++ src/components/TabButton.jsx
@@ -1,0 +1,5 @@
+const TabButton = ({children}) => {
+    return <li><button>{ children }</button></li>;
+}
+
+export default TabButton;
